% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.beta.R
\name{cdf.bet}
\alias{cdf.bet}
\title{Use Monte Carlo to estimate cdf of Beta(a, b) using R.}
\usage{
cdf.bet(x, a = 1, b = 1, n = 10000)
}
\arguments{
\item{x}{the number you want to estimate (numeric)}

\item{a}{the parameter of Beta (numeric)}

\item{b}{the parameter of Beta (numeric)}

\item{n}{the number of simulations (numeric)}
}
\value{
a number or text
}
\description{
Use Monte Carlo to estimate integration to estimate cdf of Beta(a, b).
}
\examples{
\dontrun{
cdf.bet(0.2)
cdf.bet(0.5,2,3,1e4)
}
}
