% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.cau.R
\name{cdf.cau}
\alias{cdf.cau}
\title{Use integration to estimate cdf of Cauchy distribution using R.}
\usage{
cdf.cau(y, eta1 = 0, theta1 = 1)
}
\arguments{
\item{y}{the number you want to estimate (numeric)}

\item{eta1}{the parameter of cauchy (numeric)}

\item{theta1}{the parameter of cauchy (numeric)}
}
\value{
a number
}
\description{
Use integration to estimate cdf of Cauchy distribution.
}
\examples{
\dontrun{
cdf.cau(4)
cdf.cau(2,3,5)
}
}
