% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.beta.R
\name{get.beta}
\alias{get.beta}
\title{generate a random sample from the Beta(a, b) distribution using R.}
\usage{
get.beta(n, a = 1, b = 1)
}
\arguments{
\item{n}{size of random sample (numeric)}

\item{a}{the parameter of Beta (numeric)}

\item{b}{the parameter of Beta (numeric)}
}
\value{
a random sample or a text
}
\description{
generate a random sample of size n from the Beta(a, b) distribution by the acceptance-rejection method.
}
\examples{
\dontrun{
get.beta(10,3,2)
get.beta(1e2,2,3)
a=get.beta(1000,3,2)
hist(a,probability = TRUE , main = expression(f(x)==x^(3-1)*(1-x)^(2-1)/beta(3,2)))
x=seq(0,1,0.01)
lines(x,x^(3-1)*(1-x)^(2-1)/beta(3,2))
}
}
